% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5matrix.R
\docType{methods}
\name{as.matrix,h5.matrix-method}
\alias{as.matrix,h5.matrix-method}
\title{Convert to base R matrix}
\usage{
\S4method{as.matrix}{h5.matrix}(x)
}
\arguments{
\item{x}{A h5.matrix object}
}
\description{
Extract values from a \code{h5.matrix} object
and convert to a base R matrix object
}

