% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5matrix.R
\docType{methods}
\name{dim,h5.matrix-method}
\alias{dim,h5.matrix-method}
\title{Dimensions of a h5.matrix object}
\usage{
\S4method{dim}{h5.matrix}(x)
}
\arguments{
\item{x}{A \code{h5.matrix} object}
}
\description{
Retrieve the dimensions of a \code{h5.matrix} object
}

