% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5matrix.R
\docType{methods}
\name{Extract,h5.matrix}
\alias{Extract,h5.matrix}
\alias{[,h5.matrix,ANY,ANY,logical-method}
\alias{[,h5.matrix,ANY,ANY,missing-method}
\alias{[,h5.matrix,ANY,missing,logical-method}
\alias{[,h5.matrix,ANY,missing,missing-method}
\alias{[,h5.matrix,missing,ANY,logical-method}
\alias{[,h5.matrix,missing,ANY,missing-method}
\alias{[,h5.matrix,missing,missing,logical-method}
\alias{[,h5.matrix,missing,missing,missing-method}
\alias{[<-,h5.matrix,ANY,ANY-method}
\alias{[<-,h5.matrix,ANY,missing-method}
\alias{[<-,h5.matrix,missing,ANY-method}
\alias{[<-,h5.matrix,missing,missing-method}
\title{Extract or Replace h5.matrix elements}
\usage{
\S4method{[}{h5.matrix,ANY,ANY,missing}(x, i, j, drop)

\S4method{[}{h5.matrix,ANY,ANY,logical}(x, i, j, drop)

\S4method{[}{h5.matrix,missing,ANY,missing}(x, i, j, drop)

\S4method{[}{h5.matrix,missing,ANY,logical}(x, i, j, drop)

\S4method{[}{h5.matrix,ANY,missing,missing}(x, i, j, drop)

\S4method{[}{h5.matrix,ANY,missing,logical}(x, i, j, drop)

\S4method{[}{h5.matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{h5.matrix,missing,missing,logical}(x, i, j, drop)

\S4method{[}{h5.matrix,ANY,ANY}(x, i, j) <- value

\S4method{[}{h5.matrix,missing,ANY}(x, i, j) <- value

\S4method{[}{h5.matrix,ANY,missing}(x, i, j) <- value

\S4method{[}{h5.matrix,missing,missing}(x, i, j) <- value
}
\arguments{
\item{x}{A \code{h5.matrix object}}

\item{i}{Indices specifying the rows}

\item{j}{Indices specifying the columns}

\item{drop}{Logical indication if reduce to minimum dimensions}

\item{value}{typically an array-like R object of similar class}
}

