% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5matrix.R
\docType{methods}
\name{h5.matrix}
\alias{h5.matrix}
\alias{is.h5.matrix}
\alias{is.h5.matrix,ANY-method}
\alias{is.h5.matrix,h5.matrix-method}
\title{S4 class for H5 matrices}
\usage{
h5.matrix(file, name)

is.h5.matrix(x)

\S4method{is.h5.matrix}{h5.matrix}(x)

\S4method{is.h5.matrix}{ANY}(x)
}

