\name{h5matrix-package}
\alias{h5matrix-package}
\alias{h5matrix}
\docType{package}
\title{
\packageTitle{h5matrix}
}
\description{
\packageDescription{h5matrix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{h5matrix}
\packageIndices{h5matrix}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{h5matrix}

Maintainer: \packageMaintainer{h5matrix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
