% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5matrix.R
\docType{methods}
\name{length,h5.matrix-method}
\alias{length,h5.matrix-method}
\title{Length of a h5.matrix object}
\usage{
\S4method{length}{h5.matrix}(x)
}
\arguments{
\item{x}{A \code{h5.matrix} object}
}
\description{
Get the length of a \code{h5.matrix} object
}

