% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5matrix.R
\docType{methods}
\name{ncol,h5.matrix-method}
\alias{ncol,h5.matrix-method}
\alias{nrow,h5.matrix-method}
\title{The Number of Rows/Columns of a h5.matrix}
\usage{
\S4method{ncol}{h5.matrix}(x)

\S4method{nrow}{h5.matrix}(x)
}
\arguments{
\item{x}{A h5.matrix object}
}
\value{
An integer of length 1
}
\description{
\code{nrow} and \code{ncol} return the number of
rows or columns present in a \code{h5.matrix} object.
}

