% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5matrix.R
\docType{methods}
\name{typeof,h5.matrix-method}
\alias{typeof,h5.matrix-method}
\title{The Type of a h5.matrix Object}
\usage{
\S4method{typeof}{h5.matrix}(x)
}
\arguments{
\item{x}{A \code{h5.matrix} object}
}
\description{
\code{typeof} returns the storage type of a 
\code{h5.matrix} object
}

